while i can refactor out native wrappers to a separate class
